backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2==2.9.9
pandas==2.1.4
python-dotenv==1.0.0 